requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
macholib==1.16.2
modulegraph==0.19.3
multidict==6.0.4
openai==0.27.2
py2app==0.28.5
PyYAML==6.0
requests==2.28.2
resemble==1.3.2
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
